move inject_sig. to typeutil
