+++ knowde/_feature/_shared/typeutil/__init__.py
@@ -1,2 +1,3 @@
 from .check import is_nested, is_option, is_optional  # noqa: F401
+from .func import inject_signature  # noqa: F401
 from .operate import extract_type  # noqa: F401
