+++ knowde/_feature/_shared/typeutil/func.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from inspect import Parameter, signature
+from inspect import Parameter, Signature, signature
 from typing import (
     Any,
     Callable,
@@ -41,3 +41,21 @@ def rename_argument(old: str, new: str) -> Callable[[Callable], Callable]:
         )
 
     return wrapper
+
+
+def inject_signature(
+    f: Callable,
+    t_in: list[type],
+    t_out: type | None = None,
+) -> Callable:
+    """API定義時に型情報が喪失する場合があるので、それを補う."""
+    params = signature(f).parameters.values()
+    replaced = []
+    if len(t_in) != 0:
+        for p, t in zip(params, t_in, strict=True):
+            p_new = p.replace(annotation=t)
+            replaced.append(p_new)
+    return create_function(
+        Signature(replaced, return_annotation=t_out),
+        f,
+    )
